[GM only commit]
/becomenpc will now disable tells to the target player. It will also automatically disable GM states that interfere with its functionality.

GM command /toggle will now properly turn tells on or off.

GMs will now be notified if they are ignoring tells when they first zone-in, provided their GM flag is up.

Fixed an issue with #showstats showing bogus values for size and weight. Added TellsOff variable to the output.

+++ zone/inventory.cpp
@@ -1119,7 +1119,7 @@ bool Client::SwapItem(MoveItem_Struct* move_in) {
 	}
 	// Check for No Drop Hacks
 	Mob* with = trade->With();
-	if ((with && with->IsClient() && dst_slot_id >= EmuConstants::TRADE_BEGIN && dst_slot_id <= EmuConstants::TRADE_END)
+	if ((with && with->IsClient() && !with->CastToClient()->IsBecomeNPC() && dst_slot_id >= EmuConstants::TRADE_BEGIN && dst_slot_id <= EmuConstants::TRADE_END)
 	&& GetInv().CheckNoDrop(src_slot_id)
 	&& RuleI(World, FVNoDropFlag) == 0 || RuleI(Character, MinStatusForNoDropExemptions) < Admin() && RuleI(World, FVNoDropFlag) == 2) {
 		DeleteItemInInventory(src_slot_id);
